Adding views.py
Moving ViewConfig to models.py
