+++ searchf/types.py
@@ -4,3 +4,5 @@ from typing import Tuple
 
 Size = Tuple[int, int]
 Status = str
+PaletteId = int
+ColorPair = int
